requirements.txt
llama-index-core~=0.12.17
pillow~=11.1.0
python-pptx~=1.0.2
pymilvus~=2.5.4
vanna~=0.7.6
chainlit~=2.0.5
starlette~=0.41.3
pydantic~=2.10.5
openai~=1.64.0
python-dotenv~=1.0.1
uvicorn~=0.34.0
fastapi~=0.115.6
Flask~=3.1.0
aiofiles~=23.2.1
aiohttp~=3.11.11
literalai~=0.1.103
SQLAlchemy~=2.0.37
minio~=7.2.15
PyMySQL~=1.1.1
aiomysql~=0.2.0
docling~=2.15.1
pandas~=2.2.3
huggingface-hub~=0.27.1
plotly~=6.0.0
requests~=2.32.3
sqlparse~=0.5.3
httpx~=0.28.1
autogen-agentchat
autogen-ext[openai]